size_t is overkill for rectangle dims

+++ include/visionaray/math/detail/rectangle.inl
@@ -120,14 +120,14 @@ inline T const* rectangle<xywh_layout<T>, T>::data() const
 
 template <typename T>
 MATH_FUNC
-inline T& rectangle<xywh_layout<T>, T>::operator[](size_t i)
+inline T& rectangle<xywh_layout<T>, T>::operator[](int i)
 {
     return data()[i];
 }
 
 template <typename T>
 MATH_FUNC
-inline T const& rectangle<xywh_layout<T>, T>::operator[](size_t i) const
+inline T const& rectangle<xywh_layout<T>, T>::operator[](int i) const
 {
     return data()[i];
 }
